Refactor function param type checking
